requirements-windows.txt
# SnapValid Windows 100% 可安装依赖（Python 3.10 / 3.11 通用）
pywin32>=306
PySide6>=6.6.0
pynput>=1.7.6
pyautogui>=0.9.54
pygetwindow>=0.0.9
screeninfo>=0.8.1
keyboard>=0.13.5
Pillow>=10.0.0
imagehash>=4.3.1
ddddocr>=1.4.11
numpy>=1.26.0
opencv-python>=4.8.0
pyperclip>=1.8.2
loguru>=0.6.0
